Refactoring and testing

+++ BloggingPlatform.API/Controllers/CommentController.cs
@@ -1,3 +1,4 @@
+using System.Security.Claims;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using BloggingPlatform.Business.Models.Requests;
@@ -7,6 +8,10 @@ using BloggingPlatform.Business.Models.BaseModels;
 
 namespace BloggingPlatform.API.Controllers;
 
+/// <summary>
+/// Endpoint to manage comments
+/// </summary>
+/// <param name="commentService"></param>
 [ApiController]
 [Route("api/v1/[controller]")]
 [Authorize]
@@ -22,7 +27,14 @@ public class CommentsController(ICommentService commentService) : ControllerBase
     [HttpPost]
     public async Task<IActionResult> CreateComment([FromBody] CommentRequest request)
     {
-        var response = await commentService.CreateCommentAsync(request);
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (string.IsNullOrEmpty(userId))
+        {
+            return Unauthorized("User ID not found in token.");
+        }
+        
+        var response = await commentService.CreateCommentAsync(request, userId);
+        
         return ActionResultHelper.ToActionResult(response);
     }
 
@@ -30,12 +42,17 @@ public class CommentsController(ICommentService commentService) : ControllerBase
     /// Updates an existing comment
     /// </summary>
     /// <param name="commentId">Comment ID</param>
-    /// <param name="userId">Comment ID</param>
     /// <param name="request">Comment request object</param>
     /// <returns></returns>
     [HttpPut("{commentId}")]
-    public async Task<IActionResult> UpdateComment(string commentId, string userId, [FromBody] UpdateCommentRequest request)
+    public async Task<IActionResult> UpdateComment(string commentId, [FromBody] UpdateCommentRequest request)
     {
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (string.IsNullOrEmpty(userId))
+        {
+            return Unauthorized("User ID not found in token.");
+        }
+        
         var response = await commentService.UpdateCommentAsync(commentId,  userId, request);
         
         return ActionResultHelper.ToActionResult(response);
@@ -45,11 +62,16 @@ public class CommentsController(ICommentService commentService) : ControllerBase
     /// Deletes a comment
     /// </summary>
     /// <param name="commentId">Comment ID</param>
-    /// <param name="userId">User ID</param>
     /// <returns></returns>
     [HttpDelete("{commentId}")]
-    public async Task<IActionResult> DeleteComment(string commentId, string userId)
+    public async Task<IActionResult> DeleteComment(string commentId)
     {
+        var userId = User.FindFirstValue(ClaimTypes.NameIdentifier);
+        if (string.IsNullOrEmpty(userId))
+        {
+            return Unauthorized("User ID not found in token.");
+        }
+        
         var response = await commentService.DeleteCommentAsync(commentId , userId);
         
         return ActionResultHelper.ToActionResult(response);
@@ -74,10 +96,12 @@ public class CommentsController(ICommentService commentService) : ControllerBase
     /// <param name="blogPostId">Blog post ID</param>
     /// <param name="filter">Filter object</param>
     /// <returns></returns>
-    [HttpGet("blogPost/{blogPostId}")]
+    [HttpGet("blog-post/{blogPostId}")]
     public async Task<IActionResult> GetAllCommentsByBlogPostId(string blogPostId, [FromQuery] BaseFilter filter)
     {
         var response = await commentService.GetAllCommentsByBlogPostIdAsync(blogPostId, filter);
+        
         return ActionResultHelper.ToActionResult(response);
     }
+    
 }
